Move subcommand methods to utilities
