+++ e2e/command_test.go
@@ -5,9 +5,7 @@
 package e2e
 
 import (
-	"bytes"
 	"fmt"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -261,34 +259,4 @@ func TestE2E_ReenableTaskTriggers(t *testing.T) {
 	eventCountNow := eventCount(t, dbTaskName, cts.Port())
 	require.Equal(t, eventCountBase+1, eventCountNow,
 		"event count did not increment once. task was not triggered as expected")
-}
-
-// runSubcommand runs a CTS subcommand and its arguments. If user input is
-// required for subcommand, pass it through 'input' parameter. Function returns
-// the stdout/err output and any error when executing the subcommand.
-// Note: if error returned, output will still contain any stdout/err information.
-func runSubcommand(t *testing.T, input string, subcmd ...string) (string, error) {
-	return runSubCommandWithEnvVars(t, input, []string{}, subcmd...)
-}
-
-func runSubCommandWithEnvVars(t *testing.T, input string, envVars []string, subcmd ...string) (string, error) {
-	cmd := exec.Command("consul-terraform-sync", subcmd...)
-	cmd.Env = append(cmd.Env, envVars...)
-
-	var b bytes.Buffer
-	cmd.Stdout = &b
-	cmd.Stderr = &b
-
-	stdin, err := cmd.StdinPipe()
-	require.NoError(t, err)
-
-	err = cmd.Start()
-	require.NoError(t, err)
-
-	_, err = stdin.Write([]byte(input))
-	require.NoError(t, err)
-	stdin.Close()
-
-	err = cmd.Wait()
-	return b.String(), err
 }
